Fix code according to review
