+++ react-components/src/components/Architecture/CommandButton.tsx
@@ -2,44 +2,12 @@
  * Copyright 2023 Cognite AS
  */
 
-import { type ReactElement, useState, useEffect, useMemo } from 'react';
+import { type ReactElement, useState, useEffect, useMemo, useCallback } from 'react';
 import { useRenderTarget } from '../RevealCanvas/ViewerContext';
-import { Button, Tooltip as CogsTooltip, Divider, type IconType } from '@cognite/cogs.js';
+import { Button, Tooltip as CogsTooltip, type IconType } from '@cognite/cogs.js';
 import { useTranslation } from '../i18n/I18n';
 import { type BaseCommand } from '../../architecture/base/commands/BaseCommand';
-import { OptionButton } from './OptionButton';
-import { BaseOptionCommand } from '../../architecture/base/commands/BaseOptionCommand';
 import { getButtonType, getDefaultCommand, getIcon, getTooltipPlacement } from './utilities';
-
-export const CommandButtons = ({
-  commands,
-  isHorizontal = false
-}: {
-  commands: Array<BaseCommand | undefined>;
-  isHorizontal: boolean;
-}): ReactElement => {
-  return (
-    <>
-      {commands.map(
-        (command, index): ReactElement => (
-          <CommandButtonWrapper
-            command={command}
-            isHorizontal={isHorizontal}
-            key={getKey(command, index)}
-          />
-        )
-      )}
-    </>
-  );
-};
-
-export const CreateButton = (command: BaseCommand, isHorizontal = false): ReactElement => {
-  if (command instanceof BaseOptionCommand) {
-    return <OptionButton inputCommand={command} isHorizontal={isHorizontal} />;
-  } else {
-    return <CommandButton inputCommand={command} isHorizontal={isHorizontal} />;
-  }
-};
 
 export const CommandButton = ({
   inputCommand,
@@ -58,14 +26,15 @@ export const CommandButton = ({
   const [uniqueId, setUniqueId] = useState<number>(0);
   const [icon, setIcon] = useState<IconType | undefined>(undefined);
 
+  const update = useCallback((command: BaseCommand) => {
+    setChecked(command.isChecked);
+    setEnabled(command.isEnabled);
+    setVisible(command.isVisible);
+    setUniqueId(command.uniqueId);
+    setIcon(getIcon(command));
+  }, []);
+
   useEffect(() => {
-    function update(command: BaseCommand): void {
-      setChecked(command.isChecked);
-      setEnabled(command.isEnabled);
-      setVisible(command.isVisible);
-      setUniqueId(command.uniqueId);
-      setIcon(getIcon(command));
-    }
     update(command);
     command.addEventListener(update);
     return () => {
@@ -95,25 +64,3 @@ export const CommandButton = ({
     </CogsTooltip>
   );
 };
-
-function getKey(command: BaseCommand | undefined, index: number): number {
-  if (command === undefined) {
-    return -index;
-  }
-
-  return command.uniqueId;
-}
-
-function CommandButtonWrapper({
-  command,
-  isHorizontal
-}: {
-  command: BaseCommand | undefined;
-  isHorizontal: boolean;
-}): ReactElement {
-  if (command === undefined) {
-    const direction = !isHorizontal ? 'horizontal' : 'vertical';
-    return <Divider weight="2px" length="24px" direction={direction} />;
-  }
-  return CreateButton(command);
-}
